refactor: 💡 `MainScreen` divided into other file
